Offload heartbeat to separate file to prevent circular import issue
